Add high level features for the pilot crawl.
In addition to previous simple featues based on function calls
and property accesses, we also include high-level (or complex)
features that are detected using heuristics from
the literature and aggregation of several separate function calls.

Also, we include a label for scripts who trigger HTTP requests.
We separately label script triggering third-party HTTP requests.

The exact methodology of detection for each feature are explained
in the comments.

* canvas_fingerprinting
* canvas_font_fingerprinting
* audio_context_fingerprinting
* webrtc_fingerprinting
* battery_fingerprinting
* triggers_requests
* triggers_third_party_requests